pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = [
    {name = "Gabriel Andrade"}
]
requires-python = ">=3.14"
dependencies = [
    "pytest (>=9.0.2,<10.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.17.0,<4.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-debug-toolbar (>=6.3.0,<7.0.0)",
    "psycopg2-binary (>=2.9.12,<3.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.3.1,<27.0.0)"
]
